feat: it is client.instruct now
