+++ src/SanityClient.ts
@@ -1,6 +1,5 @@
 import {lastValueFrom, Observable} from 'rxjs'
 
-import {AiClient, ObservableAiClient} from './ai/AiClient'
 import {AssetsClient, ObservableAssetsClient} from './assets/AssetsClient'
 import {defaultConfig, initConfig} from './config'
 import * as dataMethods from './data/dataMethods'
@@ -9,6 +8,7 @@ import {LiveClient} from './data/live'
 import {ObservablePatch, Patch} from './data/patch'
 import {ObservableTransaction, Transaction} from './data/transaction'
 import {DatasetsClient, ObservableDatasetsClient} from './datasets/DatasetsClient'
+import {_instruct} from './instruct/instruct'
 import {ObservableProjectsClient, ProjectsClient} from './projects/ProjectsClient'
 import type {
   Action,
@@ -25,6 +25,9 @@ import type {
   HttpRequest,
   IdentifiedSanityDocumentStub,
   InitializedClientConfig,
+  InstructAsyncInstruction,
+  InstructInstruction,
+  InstructSyncInstruction,
   MultipleActionResult,
   MultipleMutationResult,
   Mutation,
@@ -66,7 +69,6 @@ export class ObservableSanityClient {
   live: LiveClient
   projects: ObservableProjectsClient
   users: ObservableUsersClient
-  ai: ObservableAiClient
   /**
    * Private properties
    */
@@ -88,7 +90,6 @@ export class ObservableSanityClient {
     this.live = new LiveClient(this)
     this.projects = new ObservableProjectsClient(this, this.#httpRequest)
     this.users = new ObservableUsersClient(this, this.#httpRequest)
-    this.ai = new ObservableAiClient(this, this.#httpRequest)
   }
 
   /**
@@ -726,6 +727,27 @@ export class ObservableSanityClient {
   getDataUrl(operation: string, path?: string): string {
     return dataMethods._getDataUrl(this, operation, path)
   }
+
+  instruct(request: InstructAsyncInstruction): Observable<{_id: string}>
+
+  instruct<DocumentShape extends Record<string, Any>>(
+    request: InstructSyncInstruction<DocumentShape>,
+  ): Observable<IdentifiedSanityDocumentStub & DocumentShape>
+
+  /**
+   * Run an ad-hoc instruction for a target document.
+   * @param request instruction request
+   */
+  instruct<
+    DocumentShape extends Record<string, Any>,
+    Req extends InstructInstruction<DocumentShape>,
+  >(
+    request: Req,
+  ): Observable<
+    Req['async'] extends true ? {_id: string} : IdentifiedSanityDocumentStub & DocumentShape
+  > {
+    return _instruct(this, this.#httpRequest, request)
+  }
 }
 
 /** @public */
@@ -735,7 +757,6 @@ export class SanityClient {
   live: LiveClient
   projects: ProjectsClient
   users: UsersClient
-  ai: AiClient
 
   /**
    * Observable version of the Sanity client, with the same configuration as the promise-based one
@@ -763,7 +784,6 @@ export class SanityClient {
     this.live = new LiveClient(this)
     this.projects = new ProjectsClient(this, this.#httpRequest)
     this.users = new UsersClient(this, this.#httpRequest)
-    this.ai = new AiClient(this, this.#httpRequest)
 
     this.observable = new ObservableSanityClient(httpRequest, config)
   }
@@ -1431,5 +1451,26 @@ export class SanityClient {
    */
   getDataUrl(operation: string, path?: string): string {
     return dataMethods._getDataUrl(this, operation, path)
+  }
+
+  instruct(request: InstructAsyncInstruction): Promise<{_id: string}>
+
+  instruct<DocumentShape extends Record<string, Any>>(
+    request: InstructSyncInstruction<DocumentShape>,
+  ): Promise<IdentifiedSanityDocumentStub & DocumentShape>
+
+  /**
+   * Run an ad-hoc instruction for a target document.
+   * @param request instruction request
+   */
+  instruct<
+    DocumentShape extends Record<string, Any>,
+    Req extends InstructInstruction<DocumentShape>,
+  >(
+    request: Req,
+  ): Promise<
+    Req['async'] extends true ? {_id: string} : IdentifiedSanityDocumentStub & DocumentShape
+  > {
+    return lastValueFrom(_instruct(this, this.#httpRequest, request))
   }
 }
