+++ src/types.ts
@@ -1381,7 +1381,7 @@ export type {
   InstructInstructionParam,
   InstructInstructionParams,
   InstructSyncInstruction,
-} from './ai/types'
+} from './instruct/types'
 export type {
   ContentSourceMapParsedPath,
   ContentSourceMapParsedPathKeyedSegment,
